Wrapped Routes + Routes component extracted into WrappedRoutes.js
